Update schema for ClickHouse
This commit change the schema for ClickHouse to include a TTL for the
logs_local table, so that logs are deleted after 30 days.

We also improved the handling for fields which are containing, by now
explicit checking the interface if it is a number. Before this we tried
to parse every string, which lead to some false results, because labels
in Kubernetes are always strings also if they just containing a number.