Add flush interval setting (#3)
This commit introduces a new setting to define the flush interval. The
setting can be specified via the "flush_interval" property in the
config. With this new setting the entries from the buffer will now be
written to ClickHouse after the defined flush interval or when the
buffer is larger then the defined batch size setting.