Use logrus for logging and remove conf directory (#4)
We are now using the logrus package for logging, which allows us to set
the log format and have structure logging.

We also removed the conf directory and adjusted the Docker image to not
include the configuration files from it. The files are not required,
because the are mounted via a ConfigMap in Kubernetes and must always be
overwritten by users.