Add support for Kafka (#5)
It is now possible to use the Fluent Bit Kafka plugin to write the
collected logs from Fluent Bit to Kafka. We can then use the new
kafka-clickhouse application to listen for new messages in Kafka and
write them to ClickHouse.

This enables users to also use the Fluent Bit ClickHouse pipeline in
larger clusters, with a lot of nodes, because with Kafka in between of
Fluent Bit and ClickHouse we can improve the write performance to
ClickHouse.

The Docker image for the Kafka ClickHouse connector uses the
"kafka-VERSION" tag.