Make timestamp key configurable
It is now possible to configure the timestamp key for the Kafka to
ClickHouse ingestor. For that a new command-line flag
"kafka.timestamp-key" and a new environment variable
"KAFKA_TIMESTAMP_KEY" was added, which allows to customize the key,
where we should look for the correct timestamp. The default value for
this option is "@timestamp".