Rename repository to klogs (#14)
This commit finalizes the renaming process of the repository to klogs.
We decided to rename the repository, because we are not only focused on
Fluent Bit and ClickHouse anymore. Instead we are also providing an
ingester for Kafka for a while now.

We are also planning to introduce new features in the future so that
the original name doesn't make any sens.