Fix example manifests
The example manifests were not working correctly, because the Zookeeper
Pods were not create in the kind cluster because of the affinity rules.
In the ClickHouse CR the Zookeeper configuration was missing and the
Fluent Bit configuration used "Parse" instead of "Multiline.Parse" so
that the parsing didn't work correctly.