Add force number fields option
Add a new "Force_Number_Fields" option, which allows users to specify a
list of fields, which should be parsed as number. If the parsing of the
field fails it will be added to the "fields_string" list.

This is useful for fields which are only containing number but written
as string in the logs. For example the Istio access logs are containing
a field named "upstream_service_time" with values like "5", so that they
are not added to the "fields_number" list and we can not run
aggregations on it. With the new option we can add an entry
"content.upstream_service_time", so that the Fluent Bit plugin and the
ingester tries to parse this field as number and we can used it in our
aggregations.