Update create database command
We are now using the "ON CLUSTER" function of ClickHouse in the create
database so that it is not required to login into every single
ClickHouse instance to create the database. It is now enough to login to
one instance and the database is created on all other instances as well.