Call Rollback when Transaction Failed
We didn't call "tx.Rollback()" to abort a transaction, which could
caused a bug, that the plugin stopped working. This should be fixed be
calling "tx.Rollback()" via defer, so that it is always executed in case
of a failure. When the transaction was successfull it is also called,
but didn't do anything.