Use GitHub Container Registry instead of Docker Hub
Docker Hub will remove access for free organization on 14. April, so
that we need a new container registry for our Docker images. Therefor we
want to use GitHub Container Registry in the future (ghcr.io).

This commit also removes the depreacted "set-output" command and
replaces it with the recommended way from
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/